refactor: Re-arraging tables in a tree like manner
Tables with no foreign key (dependency) are put on the top level (user, currencies)

Then tables that directly depends only on the tables on the top level are placed on the second level

Then those that depend on the first and second level only are placed on the third level.

This continues until all tables are exhausted.

I am thinking for easy usability and maintainability, tables should only depend on the ones directly above them. I WILL LOOK INTO THIS.